[bar-dash-filters] Dashboard: filter charts with setBy key
[bar-dash-filters] Dashboard: filter charts with setBy key
