+++ packages/chart/src/types/ChartProps.ts
@@ -1,0 +1,7 @@
+export type BarChartProps = {
+  seriesScale: Function
+  xMax: number
+  xScale: Function
+  yMax: number
+  yScale: Function
+}
