+++ packages/dashboard/index.html
@@ -19,6 +19,7 @@
   </head>
 
   <body>
+    <div class="react-container" data-config="/examples/private/filtered-dash.json"></div>
     <!-- <div class="react-container" data-config="/examples/default.json"></div> -->
     <div class="react-container" data-config="/examples/all-components.json"></div>
     <div class="react-container" data-config="/examples/default-multi-dataset.json"></div>
